Move session state fields and methods to SessionRemote
